Refactor generated reports table data view
