+++ Clients/src/presentation/components/Table/ReportTable/TableBody/styles.ts
@@ -1,0 +1,7 @@
+export const styles = {
+  setting: {
+    position: "sticky",
+    right: 0,
+    minWidth: "50px",
+  }
+}
